- add-prometheus-metrics, fix clang analyzer warning about allocation size.
